+++ Minecraft.World/WorldGen/Sources/HellRandomLevelSource.cpp
@@ -432,7 +432,8 @@ void HellRandomLevelSource::postProcess(ChunkSource *parent, int xt, int zt)
 	pprandom->setSeed(level->getSeed());
 	__int64 xScale = pprandom->nextLong() / 2 * 2 + 1;
 	__int64 zScale = pprandom->nextLong() / 2 * 2 + 1;
-	pprandom->setSeed(((xt * xScale) + (zt * zScale)) ^ level->getSeed());
+	// 4jcraft added casts to a higher int and unsigned
+	pprandom->setSeed((((uint64_t)xt * (uint64_t)xScale) + ((uint64_t)zt * (uint64_t)zScale)) ^ level->getSeed());
 
 	netherBridgeFeature->postProcess(level, pprandom, xt, zt);
